requirements.txt
python-dateutil>=2.8.0
google-api-python-client>=2.0.0
slack-sdk>=3.0.0
python-dotenv>=0.19.0
